moved integration test to separate folder, updated pom.xml with new configuration
