+++ src/it/java/config/CucumberTest.java
@@ -1,11 +1,11 @@
-package com.epam.trainingReport.config;
+package config;
 
 import io.cucumber.junit.Cucumber;
 import io.cucumber.junit.CucumberOptions;
 import org.junit.runner.RunWith;
 
 @RunWith(Cucumber.class)
-@CucumberOptions(features = "src/test/resources/features", glue = "com/epam/trainingReport/stepDef")
+@CucumberOptions(features = "src/it/resources/features", glue = "stepDef")
 public class CucumberTest {
 
 }
