+++ src/it/resources/features/fetch_training_report_from_database.feature
