+++ src/it/resources/features/handle_training_record_request_from_message_broker.feature
